requirements.txt
pandas==2.2.3
numpy
matplotlib==3.9.2
fastapi==0.115.5
uvicorn==0.32.0
pydantic==2.9.2
huggingface_hub==0.26.2
#pymongo==4.10.1
torch==2.5.1
transformers==4.46.2
requests==2.32.3
sentencepiece==0.2.0
langchain==0.3.7
langchain_community==0.3.7
google-generativeai==0.8.3
langchain-google-genai==2.0.4
streamlit
tiktoken
requests
bs4
plotly==5.24.1
scikit-learn